Revert "add: PanelRetiroVuelto. Integración entre retiro de vuelto y retiro d…"
