+++ src/main/java/org/main/modelos/expendedor/Expendedor.java
@@ -114,12 +114,9 @@ public class Expendedor {
      * Permite extraer el vuelto de la compra moneda por moneda.
      * @return Devuelve una moneda de 100 pesos si falta por retirar vuelto, en caso contrario, devuelve <code>null</code>.
      */
-    public Moneda getVuelto() throws RetirarVacioException {
-        Moneda vuelto = depVuelto.getObjeto();
-        if (vuelto == null) {
-            throw new RetirarVacioException("No hay vuelto que retirar.");
-        }
-
+    public Moneda getVuelto() {
+        return depVuelto.getObjeto();
+    }
 
     public ArrayList<Deposito<Producto>> getDepositosProducto() {return depositosProducto;}
     public Deposito<Producto> getDepositoProducto(Catalogo producto) {
