+++ src/main/java/org/main/vistas/PanelBotones.java
@@ -1,5 +1,6 @@
 package org.main.vistas;
 import org.main.Controlador;
+import org.main.modelos.expendedor.Expendedor;
 import org.main.modelos.productos.Catalogo;
 import org.main.modelos.moneda.*;
 import javax.swing.*;
